Infra.Automation-regr_orchestration_pipe/requirements.txt
urllib3==1.25.9
requests==2.26.0
pyyaml==6.0.1
kubernetes==20.13.0
openshift==0.13.1
jenkinsapi==0.3.11
python-gl
tabulate
beautifulsoup4
